Modified datastore's models.
Added some validation logic to User and Driver models. Merged Driver and Vehicle
models into Driver. Added the capability to manage selected provinces of Driver
model. Added some clean up methods to normalize numbers in the database: now it
will all be stored using arabic alphabet; but it does not restrict the input
to that.